Track referer if package was not found
In the last few days I saw some unresolved package like "assets/css/*.css" or "&includePaths[]=" in the logs. I don't know what's going on there, so I decided to track the request referer to get some context. The stored data are used only to improve the github-linker and will be not passed on to any third party.