Add bulk endpoint (#59)
Towards our mission to use real links (see https://github.com/OctoLinker/OctoLinker/issues/277), this adds a new bulk endpoint to our service. This is needed to reduce response times and server load when prefetching links for the OctoLinker browser extension. 

The order of the response is equal to the order in request body. If a query fails or can't be resolved the api will return `null`. 


#### Request body

```json
[
    {
        "type": "registry",
        "registry": "npm",
        "target": "request"
    },
    {
        "type": "ping",
        "target": "https://nodejs.org/api/path.html"
    },
    {
        "type": "registry",
        "registry": "npm",
        "target": "package_which_does_not_exist"
    }
]
```

### Response

```json
[
	"https://github.com/request/request",
	"https://nodejs.org/api/path.html",
	null
]
```