New caching strategy
The mem package was a quite simple solution to add caching in the first place. Now with more traffic on the service a caching strategy that supports ttl is needed to maintain a healthy memory consumption. [receptacle](https://www.npmjs.com/package/receptacle) looks pretty good for this job. Btw this PR is already running in production since yesterday so it's already battle tested.