Store popular npm packages urls (#77)
Over the last 48 hours I montiered which packages are requested the most. As expected "react" is leadign the list of top npm packges. In order to save time and bandwidth I added a cache of the top 25 requested npm modules. At some point we may want to setup a cron job to validate the list, but out of scope for now in my opinion.