npm: Support lerna-style repositories
Fixes https://github.com/OctoLinker/OctoLinker/issues/535, which is
paraphrased below:

I clicked a link to the `jest-mock` package and expected to be taken to
https://github.com/facebook/jest/tree/master/packages/jest-mock.
However, I was taken to https://github.com/facebook/jest instead.

[Lerna] is growing in popularity, so this checks to see if (for example)
the `tree/master/packages/jest-mock` URL works, and go to it instead
of the repo root if possible.

[Lerna]: https://lernajs.io/